Moved slick stuff to package models.daos.slick
